Moved methods out from Program.cs Added Notifications.cs Some minor fixes on spelling and formating Changed some method names to better explain their purpose
