+++ Bank/Notifications.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Bank
+{
+    internal class Notifications
+    {
+        // Check if current ID is the same as "highest" id when logging in
+        //  If not, change notification number to amount of new IDs (New high - old)
+        // Then Store "highest" ID in transactions when logging in
+    }
+}
